Multithreading fixes with checkpointing
The primary fix here is that a pending write is now noted on new files
by updating the first_entry_id before the entry is completed. This
ensures the checkpointer doesn't proceed until the file has been
returned. Because first_entry_id isn't used for anything beyond
checkpointing progress, it doesn't matter that the entry may be
abandoned.

The second fix was changing the order of shutdown. The checkpointer
temporarily removes the files and re-adds them at the end of the
checkpoint operation. By signalling the checkpointer should shutdown
first and waiting for it to finish, we can be sure we are only waiting
on any remaining writers before reporting the shutdown success.

The last fix was a fix in the unit tests: because of the order in which
entries were being added to the shared results, it was possible for the
checkpoint function to be invoked before the entry had been added to the
in-memory collection.