Rewrote the WAL
This implementation changes to a single active file implementation. It
turns out when throwing multiple threads at a disk, writing multiple
files slows everything down at the sync stage. This WAL design
outperforms the former significantly as the number of threads increases.

This is still rough around the edges and needs a lot of documentation.