Testing change and reverting logging
I noticed a small logic bug in an edge case that I'm surprised I can't
trigger locally if it's what was causing failures on CI. In theory what
was happening is a log entry was fully written and another thread is
synchronizing the file. The other thread skips flushing the file because
its data was already synced. It then previously used the write location
not the buffer position as the indicator of how much data was considered
fsynced.

The next time to go through and synchronize the data would be considered
fully flushed so it wouldn't be flushed. As it was the last entry in the
file before being checkpointed, whatever entry wasn't flushed can't be
read by the checkpointer.

It seems like a legit edge case, but I *cannot* get it to break locally
despite my best efforts. Hopefully I can't get CI to fail after this
change too.