Added ChunkReader
Previously, reading data from the wal required that you knew how long
the chunk was. That was an oversight -- the length is stored.
Additionally, there was no way to validate the crc on randomly read
data.

ChunkReader allows validating the crc, and it also implements Read so
that it cannot read more data than the chunk originally contained.