Added streaming chunk support
While working on Sediment, I realized that I was causing one extra copy
of data when writing to the WAL than was needed, since internally the
WAL already has a buffered writer. This commit adds the ability to write
a chunk using a type that implements io::Write, provided that you know
the length of the chunk ahead of time.

The main change to support this is to move the crc from the header of
the chunk to be the tail end of the chunk instead.