Checkpointing can no longer interrupt data reads
It is now safe for ChunkReaders to exist while a checkpointing operation
is finishing up. Now, once you successfully have a ChunkReader, you can
succeed with the read, and the checkpoint operation will wait until all
readers have been dropped before allowing the file to be reused.

Additional readers are prevented from being opening after the
checkpointer finishes by renaming the file first. After renaming the
file, future readers will no longer be able to open the file because it
won't be able to be found.