Fixed issue with recovering multiple segments
When recovering segments, if two files were left uncheckpointed, the
latest file would become the new file for writing the next entry to. The
previos file(s), however, would be sent to the checkpointer.

The problem was, the checkpointer was checking if any entries had been
written, but the last_entry_id wasn't properly being updated when the
file was being scanned on load.

The other fix that wasn't able to be triggered by a unit test is if
there are three or more files left uncheckpointed -- the order was for
some reason reversed when sending to the checkpointer, which would have
yielded out of order checkpoint_to calls.