Fixed edge case when reusing log files
My previous overwrite strategy wasn't sound -- if a file was being
overwritten, and an existing entry happened to land on the boundary of
when the new entry ended, it would previously have been read.

By checking that the read entry id is greater than or equal to the entry
id that the file was created with, we can avoid the problem entirely,
and we can get rid of the extra logic when overwriting a file.