Added WriteAheadLog parameter to checkpoint_to
While working on Sediment's checkpointing logic, I realized I had a
chicken and an egg problem: I needed Sediment's LogManager to have a
reference to the WriteAheadLog, but the LogManager needed to be created
as part of the WriteAheadLog's initialization process.

While I could have solved this with an Arc<Mutex<WriteAheadLog>>, that
felt wrong. As I thought about it more, the simplest solution was to
pass in a reference to the WriteAheadLog during the checkpoint_to
function call. If Sediment potentially needs this, it seems to reason
that other users might need it as well.