Switched file operations to file-manager
As part of working my way towards integrating this into Nebari, I wanted
to unify some of the IO and syncing code between Sediment and OkayWAL.
This is the first step along that path.

The side effect is that this new file-manager crate provides an
in-memory write-ahead log. That's kind of silly, but it will be very
useful for testing, but also for allowing Sediment consumers like Nebari
(eventually) to offer an in-memory version using the same code as the
on-disk version.